+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Count duration for entries spanning multiple days for each individual day
+- Weekly aggregation in story time charts
+
+## [0.8.112] - 2022-07-19
 ### Changed:
 - Improved query for substring matched stories